+++ R/sync.R
@@ -33,10 +33,9 @@
 #'
 #'     For \strong{cv_reset} and \strong{cv_signal}: zero (invisibly).
 #'
-#' @details Pass the \sQuote{conditionVariable} to the signalling forms of the
-#'     asynchronous receive functions: \code{\link{recv_aio_signal}} or
-#'     \code{\link{request_signal}}. Alternatively, to be notified of a pipe
-#'     event, pass it to \code{\link{pipe_notify}}.
+#' @details Pass the \sQuote{conditionVariable} to the asynchronous receive
+#'     functions \code{\link{recv_aio}} or \code{\link{request}}. Alternatively,
+#'     to be notified of a pipe event, pass it to \code{\link{pipe_notify}}.
 #'
 #'     Completion of the receive or pipe event, which happens asynchronously and
 #'     independently of the main R thread, will signal the condition variable by